+++ apps/roblox_cores/urls.py
@@ -1,9 +1,9 @@
 from django.urls import path
-from apps.roblox_cores.views import (RobloxServerListView, RobloxServerDetailView, RobloxAccountListView, RobloxAccountDetailView)
+from apps.roblox_cores.views import (RobloxServerList, RobloxServerDetail, RobloxAccountList, RobloxAccountDetail)
 
 urlpatterns = [
-    path('roblox/servers/', RobloxServerListView.as_view(), name='roblox-server-list'),
-    path('roblox/servers/<int:pk>/', RobloxServerDetailView.as_view(), name='roblox-server-detail'),
-    path('roblox/accounts/', RobloxAccountListView.as_view(), name='roblox-account-list'),
-    path('roblox/accounts/<int:pk>/', RobloxAccountDetailView.as_view(), name='roblox-account-detail'),
+    path('roblox/servers/', RobloxServerList.as_view(), name='roblox-server-list'),
+    path('roblox/servers/<int:pk>/', RobloxServerDetail.as_view(), name='roblox-server-detail'),
+    path('roblox/accounts/', RobloxAccountList.as_view(), name='roblox-account-list'),
+    path('roblox/accounts/<int:pk>/', RobloxAccountDetail.as_view(), name='roblox-account-detail'),
 ]